requirements.txt
fastapi 
uvicorn[standard]
httpx
pytest
transformers
sentence-splitter 
SentencePiece
torch
pillow
stability-sdk